+++ src/TiledArray/tile_op/binary_wrapper.h
@@ -224,7 +224,7 @@ class BinaryWrapper {
                             madness::future_to_ref_t<decltype(eval_right)> r) {
       return BinaryWrapper_::operator()(l, r);
     };
-    return meta::invoke(continuation, eval_left, eval_right);
+    return detail::invoke(continuation, eval_left, eval_right);
   }
 
   /// Evaluate lazy and non-lazy tiles
@@ -249,7 +249,7 @@ class BinaryWrapper {
                                R&& r) {
       return BinaryWrapper_::operator()(l, std::forward<R>(r));
     };
-    return meta::invoke(continuation, eval_left, std::forward<R>(right));
+    return detail::invoke(continuation, eval_left, std::forward<R>(right));
   }
 
   /// Evaluate non-lazy and lazy tiles
@@ -273,7 +273,7 @@ class BinaryWrapper {
         [this](L&& l, madness::future_to_ref_t<decltype(eval_right)> r) {
           return BinaryWrapper_::operator()(std::forward<L>(l), r);
         };
-    return meta::invoke(continuation, std::forward<L>(left), eval_right);
+    return detail::invoke(continuation, std::forward<L>(left), eval_right);
   }
 
   /// Evaluate two lazy-array tiles
@@ -294,7 +294,7 @@ class BinaryWrapper {
     auto eval_left = invoke_cast(std::forward<L>(left));
     auto eval_right = invoke_cast(std::forward<R>(right));
 
-    if (perm_) return meta::invoke(op_, eval_left, eval_right, perm_);
+    if (perm_) return detail::invoke(op_, eval_left, eval_right, perm_);
 
     auto op_left = [this](eval_t<L>& _left, eval_t<R>& _right) {
       return op_.consume_left(_left, _right);
@@ -304,11 +304,11 @@ class BinaryWrapper {
     };
     // Override consumable
     if (is_consumable_tile<eval_t<L>>::value && left.is_consumable())
-      return meta::invoke(op_left, eval_left, eval_right);
+      return detail::invoke(op_left, eval_left, eval_right);
     if (is_consumable_tile<eval_t<R>>::value && right.is_consumable())
-      return meta::invoke(op_right, eval_left, eval_right);
+      return detail::invoke(op_right, eval_left, eval_right);
 
-    return meta::invoke(op_, eval_left, eval_right);
+    return detail::invoke(op_, eval_left, eval_right);
   }
 
   template <
